GraphView: renamed LogParser folder to Loghandling and moved all log dependent files to this folder.
