+++ src/ui/Loghandling/AP2DataPlotStatus.cpp
