+++ src/ui/Loghandling/AsciiLogParser.cpp
