+++ src/ui/Loghandling/AsciiLogParser.h
