+++ src/ui/Loghandling/BinLogParser.cpp
