+++ src/ui/Loghandling/BinLogParser.h
