+++ src/ui/Loghandling/ILogParser.h
