+++ src/ui/Loghandling/IParserCallback.h
