+++ src/ui/Loghandling/LogParserBase.cpp
