+++ src/ui/Loghandling/LogParserBase.h
