+++ src/ui/Loghandling/LogdataStorage.cpp
