+++ src/ui/Loghandling/TlogParser.cpp
