+++ src/ui/Loghandling/TlogParser.h
